This routine was added to save predicted arrival information in a new table define by Vernon and Pavlis a couple months ago.  This requires an extension to the css3.0 schema, so the function that does this does so cautiously.  That is, it will not fail if the "predarr" table is not known.
I also added a new table to the extensions involving predarr.  The new
table is called "stavel" and is used to set surface velocities in
km/sec for a given station/phase/vmodel.  save_predarr looks for
this table and hunts for a match against the phase and station it
is working on.  If no match is found it uses a default value for
P or S (trying hard to figure out if which is appropriate).  This
issue only really matters for the emergence angle calculation
(attribute ema in the predarr table).  The idea is that this way
one can use different surface velocities at different stations.

The computation for the other parameters in predarr is as good
as the slowness, time, and partial derivative estimates it
gets.  That is, takeoff angles and predicted slowness vectors at
each station are computed from partial derivatives and do not
assume a spherical earth model.  This generality was done in
anticipation of imminent desire for full 3D solutions.

This will be followed by related updates to dbgenloc and relocate that
implement this function.  dbgenloc has been extensively tested with
this new function.  relocate is untested at this moment.
glp