Kent did a lot of work collecting material for an Antelope reference guide and figuring out how to turn it into a Frame document.
I worked with the results of this for awhile, but was unsatisfied with
the process, so I took the raw data Kent collected and the basic ideas
and redesigned the process.  Since then, I've been attempting to
refine the results.  This check in is the intermediate results.

Basically, I dropped the use of parameter files to represent the
information for a reference guide, as I found it too cumbersome to
manipulate.  Instead, I returned to a more traditional style of
document, with commands something like html or troff.

There are three parts to this reorganization:

1) writing short perl routines which do simple parsing on a few kinds
of input, and produce framemaker mif output.

2) editing and reorganizing the various program/subroutine references
which make up the reference guides, to simplify their use and improve
their utility.  Correcting and improving the associated descriptive
info.

3) trying to design the reference booklet layout and look, so that it
can be easily printed and bound.  To this end, I've separated the
original material into three different booklets:  a user's reference
(with just command line programs), a scripting booklet (with tcl, perl
and matlab), and a c booklet (with all the c interfaces -- actually
fortran is in there, too, because I haven't figured out what to do
with it.  Perhaps it needs its own little booklet.)

The smaller booklets make it conceivable that the binding can be a
simple saddle staple rather than something dramatically more
expensive.

This is still preliminary, but it looks like finishing it will be a
project of several months of occasional work, not just a week or two.
The perl parsing routines are mostly adequate, but need a little
tweaking (or else a complete redesign which separates parsing and mif
conversion).

The reorganization is largely complete, but probably needs some fine
tuning.  Some work is necessary to ensure nothing important is left
out.  (I suppressed a lot of interfaces and a few programs as
obsolete, deprecated and/or private)

There's a lot of editing to be done to try to impose some consistency
in the usage of names, and to correct outright errors.

The layout needs some fine tuning.  I'm indefinite about generating a
table of contents and index, though framemaker makes this easy to do.
I haven't figured out yet how to get the results to look nice, though.

I'll probably be checking in changes regularly for the next few days.