Here's the reasoning:
1) Framemaker is a great language for writing extensive documentation
such as the Matlab tutorial.

2) Having the derived products, such as PDF files in their various
install locations, go out of date with respect to the source .fm
files is not optimal.

3) Therefore I built the conversion from .fm to PDF into the make
process with the Framemaker fmbatch command.

4) Not everyone who compiles the contrib code might have Framemaker
installed or installed properly, yet I don't want to sacrifice the
above benefits and don't want to go through the mess of setting up
a make config mechanism

5) Therefore I wrote a little wrapper around the Framemaker fmbatch. When
Framemaker is installed, the script runs fmbatch and also makes a
"source" file .xpdf for distribution, which is used as an alternative
compile mechanism in installations that do not have Framemaker. The intermediate
file Antelope....xpdf gets submitted to CVS in addition to the framemaker
source, and I think everybody wins.