Final implementation of S-P code.  Earlier I had fixed some flawed logic in the autoedit routine that looked for station with bad clocks. This was working.  I found, however, that I had been overly clever and held back changes to save_results that would save the results of the S-P type calculations in the output database wanting to be sure the previous part worked first.  These are now added.
Summary:  S-P code is working.  It writes an elog_log entry for
each station it forces the use of - phases on.  It will silently
truncate stations with a bad clock without one half of a
minus phase pair.  e.g. if you only have a P pick for AAK and
AAK had a bad clock a that time you will observe the mysterious behaviour
in dbloc2 that AAK's P wave will be grey after a location.
I probably ought to make that case be marked in the log file
but it requires changes to a libgenloc routine.