I left the three example files vmodel, vmodel.mod1d, and vmodel.ttcalc in the source-code directory for the tt1dcvl library. However, I removed them from the automatic install into DATADIR. I'm pretty sure it's a mistake to have the database tables themselves auto-install from this source-code directory. Otherwise, everytime I remake the library in contrib, my carefully-tailored database for Alaska gets blown away by the auto-installed example.
The alternative would be to expect everybody to have a divergent version of vmodel
in their source-code directory, keeping it permanently out-of-sync with the contrib
area and relying on the politeness of CVS not to mess up the local copy on checkout updates.
However, we already have another mechanism in Alaska to handle auto-install of local
mods [a separate CVS tree, actually]. Other institutions might want the option of
sidestepping altogether the complexity of CVS for local configuration files like vmodel*.
Thus, I think the best solution in general is not to have vmodel* auto-install from
this Makefile.

Feedback welcome