Working but intermediate version.
* Remove dbrecenteqs_setup.pf and dbrecenteqs_setup.xpl; incorporate
the functions of this executable into the main program. Intent is
to have one simple command to run that does all possible setup, and
one parameter file that's as simple as possible. Will take several
iterations.

* Started improving documentation.

* A number of parameter-file changes. Incorporate dbrecenteqs_setup.pf
into dbrecenteqs.pf. More changes will come, thus it's probably not
worth checking out and upgrading to this intermediate version.

* Added another allowed pf-to-pixfile conversion method: "imagick".

* Allow relative pathnames for some external files--logos, ancillary
databases etc., and create the new $ANTELOPE/data/dbrecenteqs. If
pathname isn't absolute or ./something, the default is to look in
$ANTELOPE/data/dbrecenteqs. This cleans up hard-wires in the
parameter file, cleans up installation, and allows things like logos etc.
to default where possible to standard-issue pieces where possible.
Keep the defaults in the Data/ subdirectory previously checked in.

* Add sophistication to the check for expansion schema. Politely
request its addition if not present.

* Add jpeg support to conversion-method "pnm". Requires pnmtojpeg,
which actually does exist.

* Allow keep_ndays to be present but set to 0, implying no cleanup of
database.

* Fix a small bug: there was only one page per evid being generated,
but on the index and global maps, little hyperlinked squares were being
made for non-preferred origins as well as preferred origins. Fixed.

* Fix Program_name assignment for error messages.