From Marina-
The dbms is an old program and it used an old API where '0' can be
sent as a buffer and it expected get allocated buffer with data filled
in.
I modified dbms to use new trgetwf API.

From Frank-

Modified to use only P waves. Previously calculated for S and P waves.