Hi Chad --
The addition of all these includes to slplatform.h allows slink2orb to compile under
Darwin.  It's possible it fouls up the compilation in solaris or linux, though
that's hard to imagine.  I'll find out tomorrow morning and revert it
if that's the case.  Otherwise, I hope this is ok; perhaps you'll have
a better approach.

-- danq