Near complete rewrite of a large fraction of this library. Long past time to check it in.  Most of this has now been given first order tests.  The main exception is the Three_Component_Seismogram portion.  Little of that part has been tested.
Primary changes in this revision:
1.  Long list of man pages added
2.  Built a generic library of interface routines to a relational
database.  Implemented a Datascope instantantiation of the generic
database handle that is an alternative interface to most
db functions.
3.  Time_Series class defined.  Has all members implemented
that form the skeleton of something useful.  Applications will
add the flesh.  This should be viewed as an alternative to the
trace library for C++ applications.
4.  Implemented a Three_Component_Seismogram object that
is treated like a special case of a time series object.
The "special case" means the samples are 3 vectors.
This is a complete change for the initial design for
this object.
5.  Added a file for spherical_coordinate functions.  This
was mostly done to handle automatic coordinate transformation
issues for a 3c seismogram object.

The Time_Series and Three_Component_Seismogram objects existed
in the previous version of this, but the design was drastically
different.  This version makes extensive use of inheritance which
greatly streamlined the implementation of some features.  Most
notably, metadata input from a database is now streamlines as a
uniform construction mechanism.  See the new manual pages if
you are interested.

Some things are still under construction here, but this is now
mostly ready for prime time.