Primary changes result from extensive testing and validation of the Three_Component_Seismogram object member functions.  If anyone else actually tried to use this stuff before, it had a lot of serious errors.  This version is no doubt not bug free, but at least has the ones out in plain sight killed.
Although the primary changes are to the Three_Component_Seismogram
objects a number of cascades came along with these changes.  They
are of two major forms:
1.  I modified the dbpp object definition to handle grouped tables
more effectively.  The most important change this required was a new
boolean that is needed to signal that a Datascope_Handle is a bundle
pointer.
2.  I discovered a fundamental misunderstanding I had about how
constructors were supposed to be done with one object as a member
of another.  Incorrect usage led to memory access errors on
constructions and duplicate free calls on destruction that are now fixed.

Finally, the man pages have been reworked some, especially the
ones on 3c seismograms, of course.

A new algorithm was added to implement Kennett's free surface
transformation matrix.