Small bug fix (was sending output to stdin -- whoops) and conversion to elog facility.
Question: I need to read a PF from a socket, do something with it, and then
output the results on the same socket.  I was hoping that pfin(stream,*Pf)
would return immediately after encountering __END__, but it seems to wait
for EOF (?).   Suggestions?  I could try shutting down half of the
connection (which I assume would cause it to indicate EOF), but I was
hoping to be able to send several parameter files in series, deliminated by
"__END__".