This fixes several problems in this library encountered in trying to get a program using the Three_Component_Ensemble database constructor to work properly: 1.  Regularized the metadata namespace.  Turned out different functions written at different times had inconsistent names for metadata (variable length header) fields.  Previous test programs didn't see this because they worked on pieces that were themselves self consistent but not consistent between themselves.  A Three_Component_Ensemble object hits both the Time_Series and Three_Component_Seismogram objects and requires consistent metadata between all of them.
2.  Fixed several problems in dbpp.cc related to grouping.  There is
a difficult problem in trying to abstract the process that Dan calls
a "bundle" used in grouping.  The Datascope_Handle object tries to
deal with this issue by keeping a parent table and using a flag to
denote that the object points to a bundle pointer.  Some things
were wrong before.  There may still be residual problems on this issue,
but at least the Three_Component_Ensemble constructor works now.

3. The seispp_attribute_maps.pf file defines some aliases for css3.0
that are regularized with the code.  A long term issue is how to
maintain the namespace in this pf with the internal namespace of
the programs in this library.