Design flaw in the Datascope Handle object concept was mercifully easy to fix.  It is the infamous problem of maintaining the integrity of integer id's in a table.
The change required here was that the Attribute_Map definitions
(seispp_attribute_maps.pf file) now has some attributes marked
as "true" that are integer id keys (e.g. evid, orid, chanid and wfid).
This is goes in parallel with a change to the Attribute_Property
object (which the Attribute_Map builds) that can now mark an attribute
as an integer key to be handled specially.

The dbsave routine for data (readwrite) now contains a line to
guaranteed the integrity of keys like wfid.  This is of minor
importance to data indexed in a wfdisc where in Datascope the
wfid key is treated as an alternate key. It was an essential
change for work in progess with a pwfdisc table that is not
dogmatic about requiring a sta:chan key.  Processed waveforms
lose a concept of any station name in some algorithms (notably
plane wave migration which is creating this need).  In the pwfdisc
table the pwfid key is THE relational key used to avoid this problem, albeit
at other costs.  Point is that the dbsave routine in libseispp was
designed to be schema neutral and will (I hope - currently not tested)
work with pwfdisc as well as wfdisc.