Fairly major changes to mesh with plane wave migration code. All made during Nov  2004 trip to Venezuela.  Changes compile on linux but are totally untested.  Unclear if this will even compile on Solaris.
Major changes of note are:
1.  added == and != operators for GCLgrid and objects that inherit
them.  Equality is defined as having congruent transformation properties
not equality of components.

2.  Some path integral related functions need to compute travel times
in 3d media needed for plane wave migration.

3.  A mapper that allows a path specified in one grid to be mapped
into the coordinate system for another.  Useful function in the
migration code as it allows a 3d model to be specified on different
grid geometries than plane wave components.

4.  update documentation

5.  Some changes to db routines that fixed some clear, undiscovered
bugs in hiding.