Several major changes precipitated by testing on plane wave migration code and related needs that arose from this: Additions: 1.  Added ascii output (overloaded << operator) for field objects. Needed and useful for looking at output with visualization software. 2.  Added copy constructors for field objects.  In c++ copy constructors are often assumed in hidden places that cause problems. They were an essential ingredient I did not have before out of ignorance.
Bug Fixes:
1.  The existing copy constructors did not use inheritance correctly.
This has been fixed.

2.  The += operator was not working correctly.  It did not handle
a common situation where a grid point was near an edge correctly.

3.  A bug in the + operator with one of the indices was fixed.

4.  A couple other indexing errors in some of the field functions.

5.  Modified the extents calculation to be more fuzzy.  The extents
of the object are not made slightly larger than the bounds of the
actual object to avoid incorrectly throwing out the corner points
due to roundoff errors.

6.  A serious bug existed in the previous version in the copy constructors
for grid objects.  They did not copy the tranformation matrix and
translation vectors which would make the copy totally wrong.

Documentation was updated, but probably needs some additions for the
new copy constructors.