Add a wphp template for generation of web-served php scripts.
Reserve xphp for possible future use on executable php scripts (parallel
to the perl extension .xpl).

This is the first version of the wphp header and may not yet be totally right.

The contrib/data/php/templates location chosen here may not be the
best place for this. Feedback welcome, Dan, I can move it if you prefer
it somewhere else.