Add suffix rule for wphp and CLEAN target for *.php.
At the moment these will have to go in every Makefile that has
wphp sources.

Also, produce has one set of fixed extensions and there doesn't appear
to be a way to use it for newly defined ones, hence the suffix rule
is somewhat hard-coded. This is open to question.