adding cmg6d response files to repository in response to Roger Hansen's request.
However, I think there are a major problem with these responses.  It
appears to me that the corner frequency for all three of these files is exactly
at the nyquist frequency for the decimated output sample rate.

For example if the input sample rate is 1000 sps, decimation factor is 5,
therefor the output sample rate is 200 sps.  Looking at the corner frequency
of the filter, it occurs at 100 Hz - the Nyquist of 200 sps.

These coeffs were from Gurlap's web site.

So CAVEAT EMPTOR all users of this device.