A rather huge change that touches pretty much every file in this library. I found I had a very inconsistent naming convention for important components of the API.  Some times I had used the _ method of separating words and sometimes I used other conventions.  I found some articles on style in object oriented languages (read Java and C++) and elected to follow a style guideline I read that suggested all classes (objects) and typedefs use a mixed upper/lower case syntax.  For example, now rather than using a construct like Basic_Time_Series the name is now BasicTimeSeries. I must agree it is easier to type.  The only residual exceptions are a few things like VelocityModel_1d that I thought were clearer with the underscore present.  Note I chose to not alter method function names in this way.  There my prejudice is that things like get_string are clearer than GetString.  Especially when they appear in constructs like   m->get_string("sta");  versus m->GetString("sta"); or   m.get_string("sta");  versus m.GetString("sta");
Judgement call, but my choice.

This change does nothing to any basic algorithm here.  The only
difference is a edit of names in these files.  In case anyone else
has an application build on seispp I'm unaware of I'm also checking
in the script FIXNAMES that I used to actually make this change.
I can be used to fix any set of source files that referenced the
previous version of seispp.  (I used it, for example, to fix dbresample
which will be checked in shortly.)  BE WARNED though that you MUST
also copy the file EDITNAMES.sed which is used by the script.  If you
look at the script you see all it does is run sed on source code files
and man pages using this file as the edit script.  The file is useful
for keeping in the repository anyway to document this name change.