Major modification of this library.  Additions should be used with caution as they have not been all completely tested yet.  In fact a major reason for this commit is that the timewindow_reader.cc (a version of a station-channel view for this library) is seg faulting and I need to check this in to try to find the problem with valgrind.
Key changes are:
1.  Missed a couple name changes in the previous checking which
aimed to regularize the naming conventions.
2.  Fixed a dead MetadataList private component in ensemble objects.
This was a relic that was not necessary after I decided to have
the ensemble classes inherit Metadata.
3.  Some simple modifications to the Ensemble objects.

Additions:
1.  A filter class which is little more than a front end to BRTT's
filtering routines.  i.e. this filter object will produce the same
results as brtt's filters, and share the same limitations.  Could
be extended, but that is where it is right now.  This class is
fairly well tested and ready for prime time.

2.  Added an important new constructor for a TimeSeriesEnsemble
that is functionally similar to a station-channel-view in the
BRTT paradigm.  A ensemble built by this constructor uses
trload_css to build a suite of traces in a fixed time window.
Metadata loaded are those in the Trace4.0 table.  Look for
future procedural routines to sort and select subsets of these
ensembles and to merge them into a ThreeComponentEnsemble object.
This thing mostly works, but as noted above I'm chasing a bug
that causes it to seg fault right now when it calls trdestroy.

3.  A new set of procedures are available in the "slice_and_dice"
file.  These allow one to cut smaller segments from a larger
time window and include an "arrival" reference frame method that
acts like trexcerpt when it is keyed on arrival times instead of
time windows.

4. A stack object has been implemented as a generalized stacker
for a TimeSeriesEnsemble.  Several methods are available including
a simple stack, median stack, and a robust estimation method.
This will be a useful general data object soon, but it's adventure
land right now.  It compiles, but has not been tested at all.
It will almost certainly break if you tried to use it.  It is a
core routine for a new array cross-correlation procedure under
construction.