Implement an Antelope sysinfo library for PHP. This adds the following 18 functions: 	my_ip 	my_hardware 	my_hostname 	my_os 	my_username 	ip2name 	name2ip 	ip 	pidexecname 	pidcmdline 	pidpwd 	pid_exists 	pidinfo 	nomem 	sysloadavg 	sysmem 	sysmounted 	statvfs
The getcwd() function was omitted since it's already in native PHP.

Move the man-pages to the now-proper section '3h'; rename them to
avoid the clunky uppercase PHP which does not match the rest of the
Antelope documentation.

Move the header files to names that align exactly with the module
names, to remove all doubt in complete compliance with the
PHP LICENSE extension-naming clause.

Improve documentation a bit; refer to the underlying C man pages
for complete elaboration on concepts (these manual pages are intended
primarily to document the PHP interface syntax, not to quote the core API
man-pages at length).

N.B. The distinction between 'Antelope interface to PHP' and
'Datascope interface to PHP' is now substantive: The former is the whole thing;
the latter is one of the component dynamic-load modules.

Installation note: Though not absolutely required, it may be cleaner to do

	cd $ANTELOPE/src/contrib/data/php
	make uninstall

before updating and installing these new changes.