Complete reorganization of this library with some major additions included at the same time.  Almost every file in the library is affected.  Some high points are: 1.  All the include files have changed to reflect an element of object oriented style I had to learn the hard way.  As the number of objects defined in a library increases it becomes essential that each major object concept be encapsulated in a include file that stands alone.  For example, there is a TimeSeries.h include file that now can be used as #include "TimeSeries.h" and it will stand alone.  Before I had an hodgepodge of things that had accumulated in seispp.h.  The new approach is that all seispp object are encapsulated in a SEISPP namespace.  This should be a cleaner approach that is more easily extended.
2.  Most, but not all .h files now have an associated .cc file that
contains the code that implements the object(s) the .h file defines.
Previously a lot of the components that made up objects were fragmented
into multiple source code files.  For example, I stupidly previously had
mixed constructors and destructors in create_destroy.C.  Now all the
components of each object definition are either defined inline in the .h
files or in a single .cc file.  This was proving to be an awful maintenance
headache that was clearly only going to get worse so I elected to clean
it up.

3.  As far as I can tell from the testing I ran the new include files
are properly self contained and have guards properly set.  This means
one can, for example, do something like
#include "ThreeComponentSeismogram.h"
and expect it to not need additional includes.  Further, all the .h
files have proper guards to prevent including the same file multiple
times.

4.  I've added new Stack and MultichannelCorrelator objects.  These
are OOP implementations of the ideas I presented at the recent AGU
meeting.  These seem stable, but the front end still has some work
before it is ready for distribution.

5.  I regularized the naming convention getting rid of old ".C" files
and converting all C++ code use the ".cc" extension.  C++ compilers
accept both, but it was proving annoying to maintain the confusing
mix.