New data object type added with some associated changes.
As the first step in conversion of the multiwavelet code to be more
workable this implements a complex time series object. This required
some related changes to other files checked in here.  Most notable is
that a "dbsave" procedure is defined that saves a complex valued
time series to an antelope database.  In theory this can work
with wfdisc, but I've unilaterally defined a new foreign data type
called "cx" to mark complex-valued data.