Major change in documentation for this package.  Converted to format required by "doxygen".  The web page for this library will now be that created by doxgen instead of ccdoc, which I used before. I made this decision for two reasons: 1) doxygen is much more heavily used than ccdoc.  All of KDE, for example, is documented with doxygen. 2) doxygen is more powerful than ccdoc.  The most interesting thing that put me over the edge is the call tree graphic that is produced for all functions it is told to parse.  The configuration I'm using here shows these even for files with incomplete documentation.
At the same time I did this several parallel changes were made:
1.  All the classes in this library are now documented.  Previously
I had a lot of holes in the documentation.
2.  I fixed a couple small errors in a couple files I found
accidentally while writing the documentation.  That is, I was trying
to remember how a few functions like correlation.cc worked and in the
process found a small lurking bug.
3.  All the .h files had to have a format change to go from ccdoc
to doxgen.
4.  I added README.doxygen, SEISPP.doxy, and  doxygen.config to the
repository.  The README file gives instructions to anyone who wants to
make a local copy of the doxygen html and latex/pdf directories.
the "doxy" file is used here to contain some file level documentation
that was better placed in s seperate file than cluttering up the
include files.  Finally, doxygen.config is the configuration file
I'm currently using to create the web documentation.  Well, the
one I'll actually use when I actually install this on the web sie
will point to an absolute path.  This configuration file creates an
html and latex directory one below this.