New program for measuring phase arrival times from teleseismic arrays using cross correlation and beam forming.  Program has a fairly sophisticated GUI using the seisw library that defines a seismic widget built on the suwigb of Seismic Unix.  It uses a package called SciPlot to produce attribute plots.
The program feeds data to the MultiChannelCorrelator object found
in libseispp.  This version is driven by a list of hypocenters
readily produced by dbe from a join of event and origin.  select
evid,orid,lat,lon,depth,origin.time
and this provide a formatted input for the program.  Eventually it
will need a more sophisticated driver, but this is currently the
working model.

There is no man page yet for this program.  Usage is
dbxcor inputdb outputdb hypolist

inputdb must point at waveform data with site and sitechan
outputdb is output and contains some extension tables defined in xcor1.0
hypolist is the dbe output described above.