Add new support for subarrays.
I've attempted to generalize the concept of a subarray.  This is turning
out to be important for large arrays using the new dbxcor program.
I chose to implement subarrays as a component of a SeismicArray object.
Subarrays can be defined through a parameter file definition and
selected by calling a new "subset" method in the SeismicArray object.

Changes in seispp_attribute_maps.pf are trivial.  Added snet to all
standard schemas to make it easier to extract that name from trace data.