Extensive changes to implement subarrays.
This set of changes mirror related changes in libseispp that
provide the core of the functionality used here.  The Xcor engine
had to undergo some fairly significant changes to use the changes
in libseispp.

Be warned that this version is likely broken.  I have it running
in test mode on linux, but there seem to be some state issues
in the gui I'm having trouble tracking down.  I'm bouncing to
Solaris to use different debugging tools and to make sure this
still works on that platform.