Assorted changes to get this program to work with S wave data, which required three component processing.  Changes here are interfacing to libseispp ThreeComponentEnsemble object and related items. Approach is to read data as full 3C object, rotate to ray coordinates if requested, and extract requested component (L,R, or T). Correlation is then performed on scalar time series data.
A parameter file change was required to implement this change.  The
most significant is the addition of a StationChannelMap description
that generalizes the idea of how to sort out multiple channel codes
for the same station.  See the http://seismo.geology.indiana.edu/~pavlis/software.html
web site for information on how this works.