Fix amplitude problems and allow autoscaling. Some fairly major changes linkd to updates in libseispp just checked in. Amplitudes are now handled correctly.  Previous versions did not do so and produced null amplitude information for the output xcorbeam table.
Perhaps more significant is a major display change related to amplitudes.
There is now an autoscale option that allows freshly read data to be
autoscaled by peak value.  A related changes is the data are now
always scaled by rms with the beam window.  This greatly improves
the graphical output compared to the previous version that was
always true amplitude.  I may ultimately make this autoscaling
optional, but for now it is always done after compuing a new beam.