Fix a time dc offset bug and an off by one error that caused random seg faults.
Previous version did not get the dc offset time correct when the
data were processed recursively.  That is, dbxcor allows one to
accumulate correlations with edits to remove bad data between
correlation runs.  When run in this mode the time offset was not
being accumulated correctly.

The other problem was related to the station name attribute display.
There was  an off by one error in the loop that posted the station
names that caused a seg fault when trying to access a -1 index for
an array.  Classic off by one error.