Add capability for running Matlab on data from a C++ program. This adds a new capability that is kind of the inverse of Kent's matlab interface.  A MatlabProcessor is a data processing object with methods to load and retrieve seismic data objects.  There is also a process command to tell matlab to process data loaded. Finally, there is a run_interactive method that is useful when running in an exploratory mode.  It basically allows typing matlab commands directly when called from a C++ program.  I've found it useful in debugging programs.  Pass some data to matlab and this method can be used explore the results graphically.
There is a man page.  Read it if you are interested in this new
funtionality.  There is also a more extensive documentation of the
API that can be found at http://seismo.geology.indiana.edu/~pavlis/software.html.
Follow the link to SEISPP and look for the documentation on the
class MatlabProcessor.

Perhaps more important to see the use of this is to look at the Examples.
Examples are incomplete.  I have only examples for looking at processed
data from ThreeComponent data objects as this is the code I've been using
this API to work on.  I've decided to put it into the repository as it
is pretty stable now.  I've been using it for about a month as a tool
for debugging my plane wave migration code where it has proven  very
useful.

Be warned the template for loading metadata attributes is untested
and I have not yet implemented the template that will allow an
ensemble to remain synchronized with matlab matrices even if the
ensemble get's sorted.