This is the initial checkin of the new orbevproc framework for doing event-oriented waveform processing. Eventually, this program will work both in the ORB and the database domains.
orbevproc will grab the same pf ORB packets that encapsulate an entire
event as produced by orbassoc, including the origin, assoc, arrival,
origerr, emodel and predarr tables. The user provides perl packages that
are evaluated through an embedded perl interpreter with a certain set of
required methods that are called ob orbevproc. For each new event,
orbevproc creates a set of perl object instances from the user-supplied
perl scripts and then calls methods on each object. An initialization
method is used to define a set of station-channel-time windows for processing.
orbevproc then goes out to find the waveforms first in an archive database.
If it cannot find everything there it then looks in an ORB. At user defined
periodic intervals, orbevproc calls a method on each processing object with
a transient traces database that has been preloaded for processing. The
perl scripts are then responsible for computing something, putting results
into a temporary output database and then reporting back to orbevproc
when the processing is complete. orbevproc then dispatches the results
according to instructions from the perl processing scripts.

orbevproc has been written in a threaded manner so each process-event
is managed asynchronously. The first example perl scripts included in
the initial release will all be related to doing magnitude computations.
Initial testing incidates that the perl scripting used in this manner
does not introduce any processing inefficiencies.

Although this is being checked into contrib now, it will not be useful
for people until the next release has become generally available (4.9).
A number of related changes needed to be made to the event associator used
by orb2dbt and some of the new features in orbassoc. Also the underlying
perlembed library went through several mods. Also, there is still a problem
with building this under Linux (sigh) related to the horrible state of
affairs in the Linux gcc world - this is becoming an all too familiar
problem. I suspect I will be able to build the perlembed stuff from
EXACTLY the same release version and libs that were used to build perl
from our Linux development machine, but you will have trouble building
it with anything else. For that reason I have included a cheat file,
ctype_b.c, that if included in the main Makefile seems to fix the problem.
Sorry for this, but it is fairly typical rinky-dink Linux bullshit.