VERY SIGNIFICANT bug fix for this library.  The previous version had frequent, very annoying scaling problems.  All but a startup transient have been fixed.  The scaling problems were solved by fixing two unrelated problems: a.  Previously the section that computed the range of start and end times did not handle dead traces correctly.  This caused occasional weird random scale changes caused by fetching a random start time. b.  After a lot of detective work I (glp) finally figured out how the seisw widget's scaling worked.  Turned out putting a ceiling on the "resolution" variables fixes a problem where the program would sometimes have an unworkable scale.
This is an interim checkin to allow a group of us to work more
effectively on dbxcor (which uses this widget) next week.
It should be considered volatile as this particular version
has two serious issues:
1)  I've left the scaffolding code for dumping window geomtry
parameters.  These aren't active, but I left them in the code
as I suspect strongly I'll need them again.
2) Right now the resolution limit parameters are hard wired.  I
have to get with Peng Wang (who wrote this widget) to understand
how to allow these to be changed with a parameter file.