Major changes to implement interface to smartpick worked out in meeting in Boulder in early June 2007.  Key changes are: 1.  Front end is fundamentally different including the command line. Front end is now intended to be driven by pairs: orid phase For example, 2114 P 2114 S 2885 P 2885 S
would aim to process P and S phases for two events (2114 and 2885).
The aim to use this interface to drive the program with smartpick
allowing simultaneous viewing of the data with dbpick and the dbxcor
gui.

2.  Back end is drastically different in two significant ways:
a.  The program now does database updates to arrival and assoc.
Previously it simply appended to a new output database and assumed
it was a different task to reconcile the new results with old.
b.  Previously the program was dogmatic about requiring schema
extension tables.  Now it simply issues a warning message if
the schema does not define these extensions.

This version should be considered unstable.  I am checking it in
to allow Frank and Kent to work on the interfacing.  Expect
numerous bug fixes in the next couple weeks as I shake this out.