Make full library byte order independent.
All of this library worked fine for both sun and intel order
machines before EXCEPT for one small segment of the three component
seismogram section.  The reason is that all other constructors
use antelope routines that have already solved this problem.
The three component seismogram object was an exception as I added
and extended datatype "3c" or (now) "c3" that is a switch for
three component data.  3c data are read as a 3xns matrix of doubles
in FORTRAN order.  now 3c implies Sparc order and c3 implies
intel order.