Display traces from two databases in dbpick.
This is used (by me) at the ANF to display seismic and soh
traces in one dbpick display, or to look at "certified"
and "non-certified" stations.

Databases undergo basic station and channel subsetting
to allow trace display in a single dbpick window.  Also
allowed is subsetting to display stations "nearby"
a given station or location.

I had a request to pass this along. Bugs still exist,
but input is welcome.